eth binary support - still using old way for now
